added

+++ Backend/src/main/java/com/dao/BookAuthorDAO.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import com.model.BookAuthor;
 @Repository
-	public interface BookAuthorDAO extends JpaRepository<BookAuthor, Integer> {
+	public interface BookAuthorDAO extends JpaRepository<BookAuthor, Long> {
 	}
